Partyholic (backend)/requirements.txt
boltons==20.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Faker==5.6.5
Flask==1.1.2
geographiclib==1.50
geopy==2.1.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pycountry==18.5.26
python-dateutil==2.8.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
text-unidecode==1.3
tqdm==4.55.1
urllib3==1.26.2
Werkzeug==1.0.1
